Fix：測定結果表示をHome Fragmentから移動
