+++ Project/app/src/main/java/com/example/project/fragment/HomeFragment.kt
@@ -6,7 +6,6 @@ import android.os.Bundle
 import android.view.View
 import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.ViewModelProviders
 import com.example.project.PagerAdapter
 import com.example.project.viewmodel.SharedViewModel
@@ -22,30 +21,15 @@ class HomeFragment : Fragment(R.layout.fragment_home) {
 
     private lateinit var model: SharedViewModel
 
-    // Homeで表示するTextViewを初期化
-    private lateinit var bodyWeightView: TextView
-    private lateinit var bodyFatPercentageView: TextView
-    private lateinit var skeletalMusclePercentageView: TextView
-    private lateinit var basalMetabolicRateView: TextView
-
     // スタイルとフォントファミリーの設定
     private var mTypeface = Typeface.create(Typeface.SANS_SERIF, Typeface.NORMAL)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        // Homeで表示するTextViewを取得
-        bodyWeightView = view.findViewById(R.id.bodyWeightView)
-        bodyFatPercentageView = view.findViewById(R.id.bodyFatPercentageView)
-        skeletalMusclePercentageView = view.findViewById(R.id.skeletalMusclePercentageView)
-        basalMetabolicRateView = view.findViewById(R.id.basalMetabolicRateView)
-
         model = activity?.run {
             ViewModelProviders.of(this)[SharedViewModel::class.java]
         } ?: throw Exception("Invalid Activity")
-
-        // viewmodelの値をtextViewに格納
-        updateView()
 
         // グラフの設定
         setupLineChart()
@@ -68,13 +52,6 @@ class HomeFragment : Fragment(R.layout.fragment_home) {
 
         /// adapterをセット
         viewPager.adapter = adapter
-    }
-
-    private fun updateView() {
-        bodyWeightView.text = model.bodyWeight.toString() + " kg"
-        bodyFatPercentageView.text = model.bodyFatPercentage.toString() + " %"
-        skeletalMusclePercentageView.text = model.skeletalMusclePercentage.toString() + " ％"
-        basalMetabolicRateView.text = model.basalMetabolicRate.toString() + " kcal"
     }
 
     // データ作成
